Convert times from UTC to local timezone.
When fetching the spot prices, there was a discrepancy with the results
because the date being requested was based off the local timezone instead
of UTC. This was causing the prices to be shown to not be the most
current ones. Switched from time ranges to a limit to circumvent this
and converted the times to local when printing to the console.